reconfigured literally everything dont even ask rn
