+++ PlatformIO/TeensyDevelopment/lib/EVT_AutoMode/EVT_AutoMode.h
@@ -1,0 +1,9 @@
+#ifndef EVT_AUTOMODE_H
+#define EVT_AUTOMODE_H
+
+#include <Arduino.h>
+
+// Autonomous mode function prototype.
+void updateAutonomousMode();
+
+#endif // EVT_AUTOMODE_H
