+++ PlatformIO/TeensyDevelopment/lib/EVT_Telemetry/EVT_Telemetry.h
@@ -1,0 +1,20 @@
+#ifndef EVT_TELEMETRY_H
+#define EVT_TELEMETRY_H
+
+#include <Arduino.h>
+#include <NativeEthernet.h>
+#include <NativeEthernetUdp.h>
+#include <string>
+
+// Global Telemetry objects and variables.
+extern EthernetUDP Udp;
+extern IPAddress ip;
+extern byte mac[];
+
+// Telemetry function prototypes.
+void setupTelemetryUDP();
+void sendTelemetry(float rpm, float vescVoltage, float odrvVoltage, float avgMotorCurrent, float odrvCurrent, float steeringAngle);
+void receiveUdp();
+void setDataFromUDP(const std::string &udpData);
+
+#endif // EVT_TELEMETRY_H
